Swap EventStoreClientAPITest <-> EventStoreClientAPIClusterTest
They were back to front in the files
